Repackaged a bunch more stuff i missed in the first commit
