+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/live/BlueAuto.java
@@ -27,46 +27,47 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.live;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import edu.ahs.robotics.hardware.sensors.ArdennesSkyStoneDetector;
+import edu.ahs.robotics.seasonrobots.Ardennes;
 import edu.ahs.robotics.util.ftc.FTCUtilities;
 
 
-@Autonomous(name = "-- AutoRed --", group = "Linear Opmode")
+@Autonomous(name = "-- AutoBlue --", group = "Linear Opmode")
 //@Disabled
-public class RedAuto extends LinearOpMode {
+public class BlueAuto extends LinearOpMode {
 
     ArdennesSkyStoneDetector detector;
+    Ardennes ardennes;
 
     @Override
     public void runOpMode() {
         FTCUtilities.setOpMode(this);
 /*
-        ArdennesSkyStoneDetector detector = new ArdennesSkyStoneDetector(false, true);
-        PartialPursuitAuto auto = new PartialPursuitAuto(false);
+        ArdennesSkyStoneDetector detector = new ArdennesSkyStoneDetector(false, false);
+        BaseAuto auto = new BaseAuto(true);
 
-        List<List<Point>> route36 = GCodeReader.openFile("red36.csv");
-        List<List<Point>> route25 = GCodeReader.openFile("red36.csv");
-        List<List<Point>> route14 = GCodeReader.openFile("red36.csv");
+        List<List<Point>> route36 = GCodeReader.openFile("foo");
+        List<List<Point>> route25 = GCodeReader.openFile("foo");
+        List<List<Point>> route14 = GCodeReader.openFile("foo");
 
-        Path quarry36 = new Path(route36.get(0), false, 0.5, new double[][]{{10, .7}, {22, .7}});
-        Path toFoundation36 = new Path(route36.get(1), false, new double[][]{{0, .5}, {12, 1}, {12, 1}}); //32
-        Path pullFoundation36 = new Path(route36.get(2), 24, 24, 48, 28,2,10, false);
-        Path quarry236 = new Path(route36.get(3), 12, 12, 20,6,2,1, false);
-        Path foundation236 = new Path(route36.get(4), 12, 12, 44,9,2,1, false);
-        Path quarry336 = new Path(route36.get(5), 12, 12, 44,9,2,1, false);
-        Path foundation336 = new Path(route36.get(6), 12, 12, 44,9,2,1, false);
+        Path quarry36 = new Path(route36.get(0), 12,12,22, false);
+        Path toFoundation36 = new Path(route36.get(1), 8,4,35, 9, 2, 1, false); //32
+        Path quarry236 = new Path(route36.get(2), 12, 12, 20,6,2,1, false);
+        Path foundation236 = new Path(route36.get(3), 12, 12, 44,9,2,1, false);
+        Path quarry336 = new Path(route36.get(4), 12, 12, 30,9,2,1, false);
+        Path foundation336 = new Path(route36.get(5), 12, 12, 40,9,2,1, false);
 
         Path quarry25 = new Path(route25.get(0), 12,12,22, false);
         Path toFoundation25 = new Path(route25.get(1), 8,4,35, 9, 2, 1, false); //32
         Path quarry225 = new Path(route25.get(2), 12, 12, 36,6,2,1, false);
         Path foundation225 = new Path(route25.get(3), 12, 12, 44,9,2,1, false);
-        Path quarry325 = new Path(route25.get(4), 12, 12, 44,9,2,1, false);
-        Path foundation325 = new Path(route25.get(5), 12, 12, 44,9,2,1, false);
+        Path quarry325 = new Path(route25.get(4), 12, 12, 30,9,2,1, false);
+        Path foundation325 = new Path(route25.get(5), 12, 12, 40,9,2,1, false);
 
         Path quarry14 = new Path(route14.get(0), 12,12,22, 18, 2,1, false);
         Path toFoundation14 = new Path(route14.get(1), 8,4,35, 9, 2, 1, false); //32
@@ -77,7 +78,6 @@ public class RedAuto extends LinearOpMode {
 
         Path quarry;
         Path toFoundation;
-        Path pullFoundation;
         Path quarry2;
         Path foundation2;
         Path quarry3;
@@ -89,7 +89,6 @@ public class RedAuto extends LinearOpMode {
         if (stoneConfiguration == ArdennesSkyStoneDetector.SkyStoneConfigurations.ONE_FOUR) {
             quarry = quarry14;
             toFoundation = toFoundation14;
-            pullFoundation = pullFoundation36;
             quarry2 = quarry214;
             foundation2 = foundation214;
             quarry3 = quarry314;
@@ -98,7 +97,6 @@ public class RedAuto extends LinearOpMode {
         } else if (stoneConfiguration == ArdennesSkyStoneDetector.SkyStoneConfigurations.TWO_FIVE) {
             quarry = quarry25;
             toFoundation = toFoundation25;
-            pullFoundation = pullFoundation36;
             quarry2 = quarry225;
             foundation2 = foundation225;
             quarry3 = quarry325;
@@ -107,7 +105,6 @@ public class RedAuto extends LinearOpMode {
         } else {
             quarry = quarry36;
             toFoundation = toFoundation36;
-            pullFoundation = pullFoundation36;
             quarry2 = quarry236;
             foundation2 = foundation236;
             quarry3 = quarry336;
@@ -115,8 +112,10 @@ public class RedAuto extends LinearOpMode {
 
         }
 
-        auto.setPaths(quarry, toFoundation, pullFoundation, quarry2, foundation2, quarry3, foundation3);
+        auto.setPaths(quarry, toFoundation, null, quarry2, foundation2, quarry3, foundation3);
+
         auto.start();
 */
     }
+
 }
